modify train_utils.py : do not remove .wav

+++ backend/ai/moment-ai/emotion2vec/train_utils.py
@@ -14,9 +14,10 @@ def extract_features(csv_file, model, label_dict):
     for wav_file, label in zip(datasets_csv['path'], datasets_csv['labels']):
         npy_file = os.path.join(DATA_PATH, os.path.splitext(wav_file)[0] + ".npy")
         
-        print(os.path.join(DATA_PATH, wav_file), os.path.exists(os.path.join(DATA_PATH, wav_file)))
-        npy_feature = model.generate(os.path.join(DATA_PATH, wav_file), granularity="utterance")
-        np.save(npy_file, npy_feature)
+        if os.path.exists(wav_file):
+            print(os.path.join(DATA_PATH, wav_file), os.path.exists(os.path.join(DATA_PATH, wav_file)))
+            npy_feature = model.generate(os.path.join(DATA_PATH, wav_file), granularity="utterance")
+            np.save(npy_file, npy_feature)
         
         datasets_list.append((npy_file, label_dict[label]))
